+++ src/main/java/uz/pdp/pdperp/service/UserService.java
@@ -4,8 +4,6 @@ package uz.pdp.pdperp.service;
 import lombok.RequiredArgsConstructor;
 
 import org.modelmapper.ModelMapper;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -17,7 +15,7 @@ import uz.pdp.pdperp.entity.UserEntity;
 import uz.pdp.pdperp.entity.enums.UserRole;
 import uz.pdp.pdperp.exception.DataAlreadyExistsException;
 import uz.pdp.pdperp.exception.DataNotFoundException;
-import uz.pdp.pdperp.exception.jwt.JwtService;
+import uz.pdp.pdperp.config.jwt.JwtService;
 import uz.pdp.pdperp.repository.UserRepository;
 
 import java.time.LocalDateTime;
@@ -30,17 +28,16 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class UserService  {
+public class UserService {
     private final ModelMapper modelMapper;
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
-
     private final JwtService jwtService;
 
     public String add(UserCreateDto dto) {
         Optional<UserEntity> userEntity = userRepository.findByUsername(dto.getUsername());
-        if(userEntity.isPresent()) {
-            throw  new DataAlreadyExistsException("User already exists");
+        if (userEntity.isPresent()) {
+            throw new DataAlreadyExistsException("User already exists");
         }
         UserEntity map = modelMapper.map(dto, UserEntity.class);
         setPermissions(map, dto.getPermissions());
@@ -49,17 +46,14 @@ public class UserService  {
         return "Successfully signed up";
     }
 
-    private UserEntity setPermissions(UserEntity userEntity, Set<String> permissions) {
-        Set<Permission> collect = permissions.stream()
-                .map(permission -> Permission.valueOf(permission.toUpperCase()))
-                .collect(Collectors.toSet());
+    public UserEntity setPermissions(UserEntity userEntity, Set<String> permissions) {
+        Set<Permission> collect = permissions.stream().map(permission -> Permission.valueOf(permission.toUpperCase())).collect(Collectors.toSet());
         userEntity.setPermissions(collect);
         return userEntity;
     }
 
     public JwtResponse signIn(AuthDto dto) {
-        UserEntity user = userRepository.findByUsername(dto.getUsername())
-                .orElseThrow(() -> new DataNotFoundException("user not found"));
+        UserEntity user = userRepository.findByUsername(dto.getUsername()).orElseThrow(() -> new DataNotFoundException("user not found"));
         if (passwordEncoder.matches(dto.getPassword(), user.getPassword())) {
             return new JwtResponse(jwtService.generateToken(user));
         }
@@ -69,28 +63,30 @@ public class UserService  {
     public List<UserEntity> getAll() {
         return userRepository.findAll();
     }
+
     public UserEntity updateRole(UUID id, String role) {
-        UserEntity userEntity = userRepository.findById(id).orElseThrow(
-                () -> new DataNotFoundException("user not found")
-        );
+        UserEntity userEntity = findById(id);
         userEntity.setRole(UserRole.valueOf(role));
-        userEntity.setUpdatedDate(LocalDateTime.now());
         return userRepository.save(userEntity);
     }
+
     public UserEntity updatePermission(UUID id, Set<Permission> permissions) {
-        UserEntity userEntity = userRepository.findById(id).orElseThrow(
-                () -> new DataNotFoundException("user not found")
-        );
+
+        UserEntity userEntity = findById(id);
         userEntity.setPermissions(permissions);
         userEntity.setUpdatedDate(LocalDateTime.now());
         return userRepository.save(userEntity);
     }
 
     public String delete(UUID userId) {
-        userRepository.findById(userId).orElseThrow(
-                () -> new DataNotFoundException("user not found")
-        );
+        if (!userRepository.existsById(userId)) {
+            throw new DataNotFoundException("user not found");
+        }
         userRepository.deleteById(userId);
         return "user delete";
+    }
+
+    private UserEntity findById(UUID userId) {
+        return userRepository.findById(userId).orElseThrow(() -> new DataNotFoundException("user not found"));
     }
 }
